set renter allowances and move GET/POST calls to new files
