+++ node/api/client/gateway.go
@@ -1,0 +1,19 @@
+package client
+
+import (
+	"github.com/NebulousLabs/Sia/modules"
+	"github.com/NebulousLabs/Sia/node/api"
+)
+
+// GatewayConnectPost uses the /gateway/connect/:address endpoint to connect to
+// the gateway at address
+func (c *Client) GatewayConnectPost(address modules.NetAddress) (err error) {
+	err = c.Post("/gateway/connect/"+string(address), "", nil)
+	return
+}
+
+// GatewayGet requests the /gateway api resource
+func (c *Client) GatewayGet() (gwg api.GatewayGET, err error) {
+	err = c.Get("/gateway", &gwg)
+	return
+}
