+++ node/api/client/renter.go
@@ -1,0 +1,19 @@
+package client
+
+import (
+	"net/url"
+	"strconv"
+
+	"github.com/NebulousLabs/Sia/modules"
+)
+
+// RenterPost uses the /renter endpoint to change the renter's allowance
+func (c *Client) RenterPost(allowance modules.Allowance) (err error) {
+	values := url.Values{}
+	values.Set("funds", allowance.Funds.String())
+	values.Set("hosts", strconv.FormatUint(allowance.Hosts, 10))
+	values.Set("period", strconv.FormatUint(uint64(allowance.Period), 10))
+	values.Set("renewwindow", strconv.FormatUint(uint64(allowance.RenewWindow), 10))
+	err = c.Post("/renter", values.Encode(), nil)
+	return
+}
